% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comat-package.R
\docType{package}
\name{comat-package}
\alias{comat}
\alias{comat-package}
\title{comat: Creates Co-Occurrence Matrices of Spatial Data}
\description{
Builds co-occurrence matrices based on spatial raster data.
    It includes creation of weighted co-occurrence matrices (wecoma) and 
    integrated co-occurrence matrices 
    (incoma; Vadivel et al. (2007) <doi:10.1016/j.patrec.2007.01.004>).
}
\author{
\strong{Maintainer}: Anoni Mize \email{anoni@mi.ze}

}
\keyword{internal}
