% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cocove.R
\name{get_cocove}
\alias{get_cocove}
\title{Create a co-located co-occurrence vector (cocove)}
\usage{
get_cocove(x, type = "ordered", normalization = "none")
}
\arguments{
\item{x}{A matrix - an output of the \code{\link[=get_cocoma]{get_cocoma()}} function}

\item{type}{The type of pairs considered.
Either "ordered" or "unordered".
The default is "ordered".}

\item{normalization}{Should the output vector be normalized?
Either "none" or "pdf".
The "pdf" option normalizes a vector to sum to one.
The default is "none".}
}
\value{
A co-located co-occurrence vector
}
\description{
Converts a co-located co-occurrence matrix (cocoma) to
a co-located co-occurrence vector (cocove)
}
\examples{
library(comat)
library(raster)
data(x, package = "comat")
data(x_na, package = "comat")
plot(x)
plot(x_na)

coom = get_cocoma(as.matrix(x), as.matrix(x_na))
coom

coov = get_cocove(coom)
coov
}
