% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wecove.R
\name{get_wecove}
\alias{get_wecove}
\title{Create a weighted cooccurrence vector (wecove)}
\usage{
get_wecove(x, ordered = TRUE)
}
\arguments{
\item{x}{A RasterStack}

\item{ordered}{The type of pairs considered.
Either ordered (TRUE) or unordered (FALSE).
The default is TRUE.}
}
\value{
A RasterStack
}
\description{
Converts a weighted cooccurrence matrix (wecoma) to
a weighted cooccurrence vector (wecove)
}
