% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comat-data.R
\docType{data}
\name{raster_w_na}
\alias{raster_w_na}
\title{A matrix with weights and missing values}
\format{A matrix}
\usage{
raster_w_na
}
\description{
A matrix with weights and missing values
}
\keyword{datasets}
