% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comat-data.R
\docType{data}
\name{raster_x_na}
\alias{raster_x_na}
\title{A matrix with categories and missing values}
\format{
A matrix
}
\usage{
data(raster_x_na)
}
\description{
A matrix with categories and missing values
}
\keyword{datasets}
