% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_wecoma}
\alias{rcpp_get_wecoma}
\title{Multiply a number by two}
\usage{
rcpp_get_wecoma(x, w, directions, fun = "mean")
}
\arguments{
\item{x}{A matrix.}

\item{w}{A matrix.}

\item{directions}{A matrix.}

\item{fun}{"mean", "geometric_mean", or "focal".}
}
\description{
Multiply a number by two
}
