% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comat-data.R
\docType{data}
\name{w_na}
\alias{w_na}
\title{A RasterLayer with weights and missing values}
\format{A RasterLayer}
\usage{
w_na
}
\description{
A RasterLayer with weights and missing values
}
\keyword{datasets}
