% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comat-data.R
\docType{data}
\name{x_na}
\alias{x_na}
\title{A RasterLayer with categories and missing values}
\format{A RasterLayer}
\usage{
x_na
}
\description{
A RasterLayer with categories and missing values
}
\keyword{datasets}
