% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comat-package.R
\docType{package}
\name{comat-package}
\alias{comat}
\alias{comat-package}
\title{comat: Creates Cooccurrence Matrices}
\description{
Builds cooccurrence matrices based on the spatial raster data.
    It includes creation of weighted cooccurrence matrices.
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)

Authors:
\itemize{
  \item Anna Dmowska
  \item Tomasz Stepinski
}

}
\keyword{internal}
