% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cocoma.R
\name{get_cocoma}
\alias{get_cocoma}
\title{Create a co-located co-occurrence matrix (cocoma)}
\usage{
get_cocoma(x, y, neighbourhood = 4)
}
\arguments{
\item{x}{A RasterLayer with categories}

\item{y}{A RasterLayer with categories}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}
}
\value{
A co-located co-occurrence matrix
}
\description{
Create a co-located co-occurrence matrix (cocoma)
}
\examples{
library(comat)
library(raster)
data(x, package = "comat")
data(x_na, package = "comat")
plot(x)
plot(x_na)

coom = get_cocoma(x, x_na)
coom

}
