% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coma.R
\name{get_coma}
\alias{get_coma}
\title{Create a co-occurrence matrix (coma)}
\usage{
get_coma(x, neighbourhood = 4)
}
\arguments{
\item{x}{A matrix with categories}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}
}
\value{
A co-occurrence matrix
}
\description{
Create a co-occurrence matrix (coma)
}
\examples{
library(comat)
library(raster)
data(x, package = "comat")
plot(x)

com = get_coma(as.matrix(x))
com

}
