% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incoma.R
\name{get_incoma}
\alias{get_incoma}
\title{Create an integrated co-occurrence matrix (wecoma)}
\usage{
get_incoma(x, neighbourhood = 4)
}
\arguments{
\item{x}{A list object containing categorical matrices with categories}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}
}
\value{
An integrated co-occurrence matrix (in form of a list of matrices)
}
\description{
Create an integrated co-occurrence matrix (wecoma)
}
\examples{
data(raster_x, package = "comat")
data(raster_w, package = "comat")
x = list(raster_x, raster_w, raster_x)

get_incoma(x)

}
